refactor utils.data module for line count linter
